Database update
Added lots of indexes to the tables and added a new layer of decision
making, as well as debugging the old code.

New formula for decision making:
least choiceState.facing aversion plus the % * 10 of target choices that
turned out badly. The formula is flawed but I like the two streams of
data